Create wrapper types for Ibft Signed messages
As future IBFT message will require only subsets of the data to be signed,
the message structures need to be modified such taht the signed-data
aspects can be a component of the message (Rather than the whole
message).

THis chnage starts this process by having a "flat" message which allows
clients to query the required round, and author of the message.